requirements.txt
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
idna==3.7
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.0
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pypdfium2==4.30.0
pytesseract==0.3.10
requests==2.32.3
rsa==4.9
urllib3==2.2.1
